at least smth works

+++ Code/lcd_functions.c
@@ -74,7 +74,7 @@ void lcd_angle(uint32_t num, char* written)
 		}
 		digit = num/devisor;
 		written[i] = digit + 0x30;
-		num -= digit*devisor;
+		num -= (uint32_t) digit*devisor;
 		i++;
 	}
 	written[i] = 0xDF;
